Use separated resolver layer for Sequel
